+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 1.6.0 (TBD)
+
+* Add a callback for link state properties in the flow frame. This allows applications to react to the
+  properties sent by the broker.
+
 ## 1.5.1 (2026-01-07)
 
 ### Bugs Fixed